model finally running
Still not sure how to add a slider to change the values of parameters. I need to rethink the way that everything is organized. Perhaps a folder for agents (maybe call it RoadUsers) and I need to think of how I am going to use property layers to make the environemnt more interactive because that is probably the most important thing here