Searching for source of Nan
right now it won't build the app because Nans are making it so that the position cannot be updated. The only plausable source is the social force model.
It's probably a zero division error if the desired direction comes out to be zero, but that probably shouldn't be happening at the first timestep.
It is time to call it for the night.