Still attempting to identify source of NaN
It seems to be in the movement calculation - the specific error is cannot convert float NaN to int.
Maybe this would all be solved by moving to a continuous space? well the int conversion would be fixed but the Nan would
still be a problem. I also want to get this finished and working before changing another foundational part
of the model. I am building unit tests and added print statements to try and find where in the model the NaN is coming from
but when you run it in Solara the print statements are not given back, just the error message. I am going to write a way to
run it in VSCode without Solara, but I need to fix some stuff with relative package imports first.