Continuous Space working
Mesa is kind of annoying. They only have an example for the experimental continuous space on their website,
and I wasn't able to find enough information about placing agents and having htem move through the regular continuous space.
For discrete space there was this .move_agents() method that was the way you moved things. The docimentation says that ContinuousSpace should
have this too, but it wasn;t working. Not sure why. I seriously need to ignore LLMs because
dear lord they just keep breaking my code. How many times do I need to learn this lesson. It's intersting that you pass the space and model to
the parent class for the experimentalCOntinuousSpaceAgents