Adding Obstacles and Surfaces
I cannot use the PropertyLayer class because that is only for discretized space. My solution for how I will implement Surfaces and Obstacles is for
them to be their own agents and to have the surface modualte the movmement parameters and the obstacles to perform pretty normally. I ran into an error
where the DataCollector was trying to access the velocity attribute of the Obstacles, which is nonsense because they do not move. I tried to fix this
by adding a new class, the MovingAgent. Bicycle and Pedestrian would then extend this class, and the DataCollector will check if each agent is an
instance of MovingAgent before storing their velocity at the current timestep.

This broke everything. The error I cannot fix is AttributeError: 'ContinuousSpace' object has no attribute 'register_agent'. It is thrown by self.model.register_agent(self)
It seems that somewhere along the way in extending classes, the space and the model get switched around. I cannot figure out where this happens.
I will add print staements to see where they are being switched around.