Added MovingAgent and Bench classes
The model finally runs again.

Obstacles are now included. They are their own file in RoadUsers. The first class in this file is bench. Right now it is just a green circle.
MovingAgent is a base class that is extended by Bicycle and Pedestrian. This created an issue with the order that arguments are passed through
the various constructors and into the model. Review *args and **kwargs and make sure that arguments are in the same order across constructors.
The MovingAgent class handles the initialization of positions and goals. Right now those things are random. In order to
get optimization going i need to handle specific gpal setting and obstacle placement because ultimately that is what is going to be optimized.

For some reason movement is really stuttery. This must have something to do with the SocialForce movement calculation. I need to look into updating
it to be more in line with Fajen and Warren anyways. Perhaps reducing the speed of the agents and increasing the simualtion step will lead to
smoother movement. Also as the simulation goes on they get moving a lot faster. Not sure why this is. The velocity also switches between val and -val
as val continues to increase.

I want to include surfaces. I think in order to get this done I need to get placing objects as a parameter and optimizing that done first.
For now the plan for Surfaces is to have them be an agent and to have them modify the movement parameters of whatever is on them. I need to make use
of the get_agents_in_radius methods and get_k_nearest_agents methods.