The optimizer is running.
Right now each particle contains a dictionary of agent placements for initializing a model. Right now they are all converging on the same goal
because the .init_goal() method in MovingAgent has np.random.seed(0). I am not sure how to get the optimizer to terminate, there is no convergence
logic. Also currently there is not a great way to get data out of the optimizer.

Currently the dimensions of the optimization space = n_agents * 2. This is not great. Is there a way to diminsh this that would preserve the
relationship between moving the particle and moving each of the agents?

There is potential for a moajor refactor in the way that the model instances / agents are initialized. I'm pretty sure that persisting model
instances across timesteps in the optimizer is the most efficient way to do this. Definitely a question for another day. For the time being
the environment that corresponds to the particle will be set by PedestrianBicycleModel.set_environemnt(). It took a while for me to figure out
how to call the COntinuousSpaceAgent.remove() method on the Pedestrians and Bicycles in the model. **check if obstacle calls .remove() from CSA**

There is some serious restructuring / troubleshooting that needs to happen with filepaths so that I can run things from outside of main. Maybe put all
of the other folders in src. Right now I don't think that placing anything in visualization would be able to access the rest of the project strucutre.
Perhaps I need to learn about relative imports.

I still need to rewrite social force to have a term for angular momentum a la Fajen and Warren. I want to do that next, then surfaces after that.
Then I will have the surfaces and movement attributes interact, then I will introduce decsiion policies for obeying traffic laws. Maybe a refactor
will be built into all of that.